% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gvp.R
\name{calculate_gvp}
\alias{calculate_gvp}
\title{GVP}
\usage{
calculate_gvp(
  inputdirectory,
  outputdirectory,
  aggregated = F,
  magedef = "1sd",
  congan = 1,
  saveplot = T
)
}
\arguments{
\item{inputdirectory}{path to folder containing cleaned files. If data is pre-aggregated then use the full file path including file name of this file.
Preferred csv format but can read in others.}

\item{outputdirectory}{path to folder where output files will be stored}

\item{aggregated}{TRUE/FALSE dictates if data is pre-aggregated or in individual separate files (usual for clinical trials)}

\item{magedef}{Defining the threshold used in MAGE calculation.  MAGE is an arithmetic average of either the upward or downward
of all glycemic excursions exceeding the threshold (standard deviation of blood glucose obtained from all blood glucose
concentrations within 24-hour period). Default is 1 standarddevation ("1sd"), options are 1.5 SD ("1.5sd") , 2 SD ("2sd")
or other can be specifed as a numeric}

\item{congan}{Specificing the n number of hours in CONGA(n). Default is the numeric 1. CONGA(n) represents the SD
of all valid differences between a current observation and an observation (n) hours earlier}

\item{saveplot}{A TRUE/FALSE to save all associated plots for each patient as in a combined PDF. Default to TRUE. If FALSE no PDFs will be generated.}
}
\value{
Returns a csv files of GVP for each individual.
}
\description{
This function enables generation Glycemic Variability Percentage (GVP) from CGM data. CGM data must be cleaned and in the format as described in the README also see \code{\link[CGMprocessing:cleanCGM]{CGMprocessing::cleanCGM()}}.
This is based on the logic from Peyser 2018. The length of the CGM temporal trace over a given interval of time depends on the degree of glycemic variability: temporal traces with high glycemic variability
have greater lengths than traces with low glycemic variability.
Mathematically, this is similar to the problem posed by Mandelbrot regarding the proper measurement of the coastline of Great Britain or other coastal areas with a high degree of tortuosity.
}
\seealso{
CGMprocessing cleanCGM CHAOSindex
}
\author{
Alice Carr (University of Alberta)
}
