% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALJC_CHAOSpackageskeleton.R
\name{CHAOSindex}
\alias{CHAOSindex}
\title{Generate CHAOS index}
\usage{
CHAOSindex(
  inputdirectory,
  outputdirectory = "output",
  maxhorizon = 90,
  saveplot = T
)
}
\arguments{
\item{inputdirectory}{A string file path to a file of raw CGM downloads.}

\item{outputdirectory}{A string file path to the output data file, folder will be generated if it doesn't exist. Default folder name: "output".}

\item{maxhorizon}{A numeric of the number of minutes for the maximum horizon we want to predict to in addition to the 30 minute horizon. Default to 90 minutes.
keep as default.}

\item{saveplot}{A TRUE/FALSE to save all associated plots for each patient as in a combined PDF. Default to TRUE. If FALSE no PDFs will be generated.}
}
\value{
Description of the object returned by the function.
}
\description{
Enables generation of the CHAOS index from CGM data.
This is an concept index in development at the University of Alberta and University of Cardiff
being trialed to rapidly and effectively identify and monitor high risk individuals with
Type 1 diabetes requiring advanced treatment.
}
\author{
Alice Carr, Peter Taylor
}
