% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALJC_CHAOSpackageskeleton.R
\name{CHAOSindex}
\alias{CHAOSindex}
\title{CHAOSindex}
\usage{
CHAOSindex(
  inputdirectory,
  outputdirectory = "output",
  aggregated = F,
  maxhorizon = 90,
  saveplot = T
)
}
\arguments{
\item{inputdirectory}{path to folder containing cleaned files. If data is pre-aggregated then use the full file path including file name of this file.
Preferred csv format but can read in others.}

\item{outputdirectory}{path to folder where output files will be stored}

\item{aggregated}{TRUE/FALSE dictates if data is pre-aggregated or in individual separate files (usual for clinical trials)}

\item{maxhorizon}{A numeric value in minutes of the maximum horizon time you want to predict up until. Default is 90 (minutes).}

\item{saveplot}{A TRUE/FALSE to save all associated plots for each patient as in a combined PDF. Default to TRUE. If FALSE no PDFs will be generated.}
}
\value{
Returns PDF of CGM traces and modelling prediction windows as well as csv files of the ARMIA model "accuracy" parameters, the Mean Absolute Percentage Errors (MAPE) for a 30 minute prediction window
as well as the chosen maximum prediction horizon (default to 90 min). Also outputs the file of the real values and predicted values from this prediction for 30 mins and the chosen maximum horizon.
}
\description{
This function Enables generation of the CHAOS index from CGM data. CGM data must be cleaned and in the format as described in the README also see cleanCGM from the CGMprocessing package.
This is an concept index in development at the University of Alberta and University of Cardiff being trialed to rapidly and effectively identify and monitor high risk individuals with
Type 1 diabetes requiring advanced treatment. This package is in development.
}
\seealso{
CGMprocessing cleanCGM
}
\author{
Alice Carr (University of Alberta), Peter Taylor (University of Cardiff), Steph Hanna (University of Cardiff)
}
